+++ src/main/java/com/LetMeDoWith/LetMeDoWith/provider/OidcIdTokenProvider.java
@@ -9,7 +9,12 @@ import com.LetMeDoWith.LetMeDoWith.exception.OidcIdTokenPublicKeyNotFoundExcepti
 import com.LetMeDoWith.LetMeDoWith.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.service.SocialProviderAuthFactory;
 import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwt;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.SignatureException;
 import java.security.Key;
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
@@ -28,13 +33,11 @@ public class OidcIdTokenProvider {
     
     private final SocialProviderAuthFactory socialProviderAuthFactory;
     private final AuthTokenProvider authTokenProvider;
-    
+    private final String KID = "kid";
     @Value("${auth.oidc.aud.kakao}")
     private String KAKAO_AUD;
-    
     @Value("${auth.oidc.aud.google}")
     private String GOOGLE_AUD;
-    
     @Value("${auth.oidc.aud.apple}")
     private String APPLE_AUD;
     
@@ -49,9 +52,9 @@ public class OidcIdTokenProvider {
         AuthClient client = socialProviderAuthFactory.getClient(provider);
         OidcPublicKeyResDto publicKeyList = client.getPublicKeyList().block();
         String aud = getAudValueForProvider(provider);
-        String kid = authTokenProvider.getKidFromUnsignedTokenHeader(token,
-                                                                     aud,
-                                                                     provider.getIssUrl());
+        String kid = getKidFromUnsignedTokenHeader(token,
+                                                   aud,
+                                                   provider.getIssUrl());
         
         try {
             
@@ -92,6 +95,60 @@ public class OidcIdTokenProvider {
                 throw new RestApiException(FailResponseStatus.OIDC_ID_TOKEN_PUBKEY_NOT_FOUND);
             }
         }
+    }
+    
+    /**
+     * Unsigned token에서 kid를 가져온다,
+     *
+     * @param token Unsigned ID token
+     * @param aud   자격증명 제공자에서 발급한 어플리케이션 아이디
+     * @param iss   자걱증명 제공자의 url
+     * @return ID token의 kid 값
+     */
+    public String getKidFromUnsignedTokenHeader(String token, String aud, String iss) {
+        return (String) getUnsignedTokenClaims(token, aud, iss).getHeader().get(KID);
+    }
+    
+    /**
+     * OIDC ID Token의 iss, aud, 만료일자를 검증하고, parse한다.
+     *
+     * @param token ID Token. Signature가 제거되있어야 한다.
+     * @param aud   자격증명 제공자에서 발급한 어플리케이션 아이디
+     * @param iss   자걱증명 제공자의 url
+     * @return parsing된 ID Token.
+     */
+    private Jwt<Header, Claims> getUnsignedTokenClaims(String token, String aud, String iss) {
+        try {
+            return Jwts.parserBuilder()
+                       .requireAudience(aud)  // aud(provider에 지정한 어플리케이션 아이디) 가 같은지 확인
+                       .requireIssuer(iss) // iss(issuer)가 알맞은 provider인지 확인
+                       .build()
+                       .parseClaimsJwt(getUnsignedToken(token));
+        } catch (ExpiredJwtException e) { // 파싱하면서 만료된 토큰인지 확인.
+            throw new RestApiException(FailResponseStatus.TOKEN_EXPIRED, e);
+        } catch (SignatureException e) {
+            throw new RestApiException(FailResponseStatus.INVALID_TOKEN, e);
+        } catch (Exception e) {
+            log.error(e.toString());
+            throw new RestApiException(FailResponseStatus.INVALID_TOKEN, e);
+        }
+    }
+    
+    /**
+     * Encoding된 Signed-token 에서 signature를 분리한다.
+     *
+     * @param signedToken signature를 포함하는 token 전문
+     * @return signature가 제거된 token
+     */
+    
+    private String getUnsignedToken(String signedToken) {
+        String[] splitToken = signedToken.split("\\.");
+        
+        if (splitToken.length != 3) {
+            throw new IllegalArgumentException("올바르지 않은 Token입니다!");
+        }
+        
+        return splitToken[0] + '.' + splitToken[1] + '.';
     }
     
     private String getAudValueForProvider(SocialProvider provider) {
